reorganizing generalized dashboard builder
